speedtest-exporter: Restructure directory to golang best practices
Directory layout is according to:
https://github.com/golang-standards/project-layout
