+++ .github/workflows/build-speedtest-exporter.yaml
@@ -9,7 +9,6 @@ on:
     paths:
       - "apps/speedtest-exporter/Dockerfile"
       - "apps/speedtest-exporter/Dockerfile.cli"
-      - "apps/speedtest-exporter/*.go"
       - "apps/speedtest-exporter/go.mod"
       - "apps/speedtest-exporter/go.sum"
       - "apps/speedtest-exporter/**/*.go"
@@ -18,7 +17,6 @@ on:
     paths:
       - "apps/speedtest-exporter/Dockerfile"
       - "apps/speedtest-exporter/Dockerfile.cli"
-      - "apps/speedtest-exporter/*.go"
       - "apps/speedtest-exporter/go.mod"
       - "apps/speedtest-exporter/go.sum"
       - "apps/speedtest-exporter/**/*.go"
