+++ apps/speedtest-exporter/Makefile
@@ -19,7 +19,10 @@ build-cli:
 
 go-build:
 	make -p bin
-	go build $(GO_BUILD_FLAGS) -o bin/speedtest-exporter
+	go build $(GO_BUILD_FLAGS) -o bin/speedtest-exporter ./cmd/
+
+go-test:
+	go test -v ./...
 
 .PHONY: \
 	default \
@@ -27,4 +30,5 @@ go-build:
 	build-slim \
 	build-cli \
 	go-build \
+	go-test \
 	$(NULL)
