+++ apps/speedtest-exporter/cmd/cmd.go
