+++ apps/speedtest-exporter/cmd/cmd_test.go
