+++ apps/speedtest-exporter/pkg/collector/collector.go
