+++ apps/speedtest-exporter/pkg/collector/collector_test.go
