+++ apps/speedtest-exporter/pkg/collector/types.go
