+++ apps/speedtest-exporter/pkg/collector/types_test.go
