+++ apps/speedtest-exporter/pkg/speedtest/result-json.go
