+++ apps/speedtest-exporter/pkg/speedtest/testdata/speedtest-cli-result.json
