+++ apps/speedtest-exporter/pkg/speedtest/testdata/speedtest-cli.sh
