+++ apps/speedtest-exporter/pkg/speedtest/util.go
